+++ src/migrations/v106.ts
@@ -1,0 +1,25 @@
+// in V1.0.6 we change from hpMode toggleing between NUM and BAR to a boolean hpBar
+
+import OBR, { isImage } from "@owlbear-rodeo/sdk";
+import { characterMetadata } from "../helper/variables.ts";
+
+export const migrate105To106 = async () => {
+    console.log("Migration from 105 to 106 running");
+    await OBR.scene.items.updateItems(isImage, (images) => {
+        images.forEach((image) => {
+            if (characterMetadata in image.metadata) {
+                const metadata: Object = image.metadata[characterMetadata] as Object;
+                if (metadata.hasOwnProperty("hpMode")) {
+                    // @ts-ignore
+                    const currentMode = metadata.hpMode;
+                    const hpBar = currentMode === "BAR";
+                    // @ts-ignore
+                    metadata.hpBar = hpBar;
+                    // @ts-ignore
+                    delete metadata.hpMode;
+                    image.metadata[characterMetadata] = { ...metadata };
+                }
+            }
+        });
+    });
+};