+++ CHANGELOG.md
@@ -33,3 +33,9 @@
   - mouse drag can no longer change values (thank google)
 - buttons have proper cursor icon
 - Fix layout not working for mobile
+
+# 1.0.6
+
+- Improve HP Bar handling and make icons easier to understand
+- Add Highlighted Character to Playerview when HP Bar is active
+- Removes changing values with mouse wheel due to bug with large numbers of items